# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "This is a sample recipes API. You can find out more about the API at https://github.com/vijeyash1/gin",
    "title": "Recipes API",
    "contact": {
      "name": "Vijeyash",
      "url": "http://vijeyash.com +918489635967",
      "email": "avijeyash@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:9000",
  "basePath": "/",
  "paths": {
    "/recipes": {
      "get": {
        "description": "Returns list of recipes",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "listRecipes",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "description": "Adds a new recipe from post request",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "AddNewRecipes",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/recipes/{id}": {
      "get": {
        "description": "Search recipes based on id",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "GetOneRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "recipe id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "put": {
        "description": "Update an Excisting recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "UpdateRecipes",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull operation StatusOk"
          },
          "400": {
            "description": "invalid input StatusBadRequest"
          },
          "404": {
            "description": "invalid recipe id StatusNotFound"
          }
        }
      },
      "delete": {
        "description": "Deletes an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "DeleteRecipes",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull Operation StatusOk"
          },
          "400": {
            "description": "invalid input StatusBadRequest"
          }
        }
      }
    }
  }
}